+++ Automate/Framework/Machines/DataBasedMachine.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Linq;
 using Microsoft.Xna.Framework;
 using StardewValley;
@@ -9,39 +10,89 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines
     internal class DataBasedMachine : GenericObjectMachine<SObject>
     {
         /*********
+        ** Fields
+        *********/
+        /// <summary>The minimum machine processing time in minutes for which to apply fairy dust.</summary>
+        private readonly Func<int> MinMinutesForFairyDust;
+
+
+        /*********
         ** Public methods
         *********/
         /// <summary>Construct an instance.</summary>
         /// <param name="machine">The underlying machine.</param>
         /// <param name="location">The location containing the machine.</param>
         /// <param name="tile">The tile covered by the machine.</param>
-        public DataBasedMachine(SObject machine, GameLocation location, Vector2 tile)
-            : base(machine, location, tile, DataBasedMachine.GetMachineId(machine.Name)) { }
+        /// <param name="minMinutesForFairyDust">The minimum machine processing time in minutes for which to apply fairy dust.</param>
+        public DataBasedMachine(SObject machine, GameLocation location, Vector2 tile, Func<int> minMinutesForFairyDust)
+            : base(machine, location, tile, DataBasedMachine.GetMachineId(machine.Name))
+        {
+            this.MinMinutesForFairyDust = minMinutesForFairyDust;
+        }
 
         /// <inheritdoc />
         public override bool SetInput(IStorage input)
         {
-            if (!this.Machine.HasContextTag("machine_input"))
+            SObject machine = this.Machine;
+
+            // skip if no input needed
+            if (!machine.HasContextTag("machine_input"))
                 return false;
 
-            foreach (IContainer container in input.InputContainers)
+            // add machine input
+            bool addedInput = false;
+            foreach (IContainer container in input.OutputContainers)
             {
-                if (this.Machine.AttemptAutoLoad(container.Inventory, Game1.player))
-                    return true;
+                if (machine.AttemptAutoLoad(container.Inventory, Game1.player))
+                {
+                    addedInput = true;
+                    break;
+                }
             }
 
-            return false;
+            // apply fairy dust
+            if (addedInput)
+                this.TryApplyFairyDust(input);
+
+            return addedInput;
+        }
+
+        /// <summary>Get a machine ID for a machine item.</summary>
+        /// <param name="name">The machine's internal item.</param>
+        public static string GetMachineId(string name)
+        {
+            return new string(name.Where(char.IsLetterOrDigit).ToArray());
         }
 
 
         /*********
         ** Private methods
         *********/
-        /// <summary>Get a machine ID for a machine item.</summary>
-        /// <param name="name">The machine's internal item.</param>
-        public static string GetMachineId(string name)
+        /// <summary>Apply fairy dust from the given containers if needed.</summary>
+        /// <param name="input">The input to search for containers.</param>
+        private void TryApplyFairyDust(IStorage input)
         {
-            return new string(name.Where(char.IsLetterOrDigit).ToArray());
+            SObject machine = this.Machine;
+            int minMinutes = Math.Max(10, this.MinMinutesForFairyDust());
+
+            if (machine.MinutesUntilReady < minMinutes || !machine.TryApplyFairyDust(probe: true))
+                return;
+
+            int maxToApply = 3;
+            foreach (IContainer container in input.OutputContainers)
+            {
+                while (maxToApply > 0 && container.Inventory.ContainsId("(O)872"))
+                {
+                    if (!machine.TryApplyFairyDust())
+                        return;
+
+                    container.Inventory.ReduceId("(O)872", 1);
+                    maxToApply--;
+
+                    if (machine.MinutesUntilReady < minMinutes || !machine.TryApplyFairyDust(probe: true))
+                        return;
+                }
+            }
         }
     }
 }